+++ include/SQLiteCpp/Backup.h
@@ -27,16 +27,8 @@ namespace SQLite
  * A Backup object is used to backup a source database file to a destination database file
  * in a safe and online way.
  *
- * Resource Acquisition Is Initialization (RAII) means that the Backup Resource
- * is allocated in the constructor and released in the destructor, so that there is
- * no need to worry about memory management or the validity of the underlying SQLite Backup.
- *
- * Thread-safety: a Backup object shall not be shared by multiple threads, because :
- * 1) in the SQLite "Thread Safe" mode, "SQLite can be safely used by multiple threads
- *    provided that no single database connection is used simultaneously in two or more threads."
- * 2) the SQLite "Serialized" mode is not supported by SQLiteC++,
- *    because of the way it shares the underling SQLite precompiled statement
- *    in a custom shared pointer (See the inner class "Statement::Ptr").
+ * See also the a reference implementation of live backup taken from the official site:
+ * https://www.sqlite.org/backup.html
  */
 class Backup
 {